Specialize update_existing_value for sum_op
Hardware atomic addition is much faster than atomic addition via CAS,
since it can be performed in L2. This change specializes the implementation
for update_existing_value for the sum_op aggregator for int32, int64,
fp32, and fp64.

In my test on a GP102, the time to compute a groupby mean
on 10M entries with only 10 distinct groups improved from
~3.7s to ~60ms.

[Helps a lot with #533.]