Add `__len__` and `nbytes` to `Buffer` for Dask
Dask tries to figure out the length and/or number of bytes in a given
object before sending the object itself so the receiver can allocate
some space to fit this data. It does this by using `len(...)` and
checking the `.nbytes` property. Since we are trying to now send
`Buffer`s direct through Dask. Make sure they also have these methods so
that Dask is able to inspect their size before sending them.