Use existing `Buffer`s for serialization
As we already have `Buffer`s allocated with the relevant data, just use
them in serialization instead of creating new ones. Also it is worth
noting offsets come first and then strings. However deserialization
expects the opposite order. So make sure to reverse the children before
adding them to the frames to match that expected order for
deserialization.