Type `CUDARuntimeError`'s `status`
Since we know `status` should be a `cudaError_t` value and we can
enforce this in the constructor (as Python `int`s will be coerced to it
as needed), go ahead and do that once at the outset. Should be cheaper
to do once and then easy to reuse in simple/fast C++ code later.