Add the `Serializable` abstract base class
This provides a generic interface for handling serialization in Python.
Handles serialization with Dask through the `"cuda"` and `"dask"`
serializers. Also implements pickle serialization using `__reduce_ex__`.
For Python versions with support for pickle's protocol 5, the class also
supports out-of-band buffers for more efficient serialization.
Subclasses are responsible for implementing `serialize` and
`deserialize` to/from a Python `dict` with the `header` and a collection
of `frames`. The abstract base class handles all other serialization
using these two methods.