Fix pickling sizeof test
This test assumes that pickling the data will yield something of the
same size. While that seems reasonable, it's not quite right as a bunch
of information about how to extract the pickled data and rebuild Python
objects will be included. So instead merely check that the pickled data
is at least the size of the object. IOW no data was lost by pickling.
However don't tightly enforce equality.