Avoid chaining `Buffer`s
Previously if we constructed a `Buffer` from a `Buffer`, the new
`Buffer` would reference the old `Buffer` used to create that. However
this can result in `Buffer`s that have chained references, which is
inefficient and unneeded. Instead just shortcut around the provided
`Buffer` and reference the object owning the data. This will avoid
chained `Buffer`s and ensure the underlying data remains valid even if
the old `Buffer` scopes out.