Add minmax reduction (#6247)
Adds a reduction function to compute minimum and maximum reductions  in a single-pass. Currently doesn't support complex types like dictionaries and fixed-point due to DeviceMin and DeviceMax not supporting fixed-point.

Added tests to verify min and max return the same results as minmax. Also added a test for a column full of nulls.
closes #5751

Co-authored-by: Mike Wilson <mwilson@nvidia.com>
Co-authored-by: Jake Hemstad <jhemstad@nvidia.com>
Co-authored-by: Karthikeyan Natarajan <karthikeyann@users.noreply.github.com>
Co-authored-by: Karthikeyan <6488848+karthikeyann@users.noreply.github.com>