Fix JNI native dependency load order (#6617)
This changes the JNI native dependency load order to always load libcudf_base.so before loading libcudf_comms.so.

When built with some toolchains, libcudf_comms.so has an explicit dependency on libcudf_base.so whereas with other toolchains it does not. The JNI code currently loads libcudf_base.so and libcudf_comms.so in parallel, but this only works when building with a toolchain smart enough to realize libcudf_comms.so does not need libcudf_base.so despite linking against it.