Add strings::contains API with target column parameter (#6598)
This adds a new `cudf::strings::contains()` API that accepts a target column instead of a scalar. Each row of the target is check against the corresponding row in the source column. If the target string appears inside the source string then `true` is set in the output column for that row. 
The gtests in `find_tests.cpp` were also refactored to make it easier to add new tests and features.
This also includes an update to the `cudf str.contains()` API so that it accepts a column for the `pat` argument. An appropriate pytest was also added to `test_string.py`.