Remove 2nd type-dispatcher call from cudf::reduce for simple operations (#6727)
This PR removes the 2nd type-dispatch from the simple reduce operations: any(), all(), min(), max(), sum(), product(), and sum_of_squares(). It is replaced with logic using either the same type as the input column type or a hardcoded type as follows: any() and all() expect only BOOL8 as the output type; min() and max() expect the output type to match the column input type; sum(), product(), and sum_of_squares() allow any output type but may incur an extra scalar cast operation.
The cast operation is only required if the output type does not match the input type or if the output type is not INT64 for integer columns or FLOAT64 for floating point columns. The extra cast only applies to sum(), product(), and sum_of_squares() ops.