[REVIEW] Optimization and nested type support for contiguous_split. (#6615)
This PR adds support for list and struct columns to contiguous_split. It also fundamentally reworks how the code operates so that the number of memory allocations and kernel calls done is a constant. Previously the number of kernel calls we were doing was proportional to the # of columns * the # of splits which could cause serious slowdowns.