fix excluding cufile tests by default(#6988)
I think this is how I started and verified it was working, but then I was trying to exclude the source as well, which didn't work for tests. Then I realized we need the source built for the plugin so remove it. Anyway, no presubmit CI check is really painful. :(

@revans2 

```console
$ mvn test
...
[WARNING] Tests run: 775, Failures: 0, Errors: 0, Skipped: 4

$ mvn test -DUSE-GDS=ON
...
[WARNING] Tests run: 777, Failures: 0, Errors: 0, Skipped: 4
```

Authors:
  - Rong Ou <rong.ou@gmail.com>

Approvers:
  - Robert (Bobby) Evans

URL: https://github.com/rapidsai/cudf/pull/6988