Fix thrust failure when transfering data from device_vector to host_vector with vectors of size 1 (#7382)
This is a fix for the problem that popped up in the bug https://github.com/rapidsai/cudf/issues/6364.

I debug this problem and discovered that thrust crashes only when the vector size is 1. There should not be any problem with the vector size, thus the bug should be due to something else. Finally, I found a fix for this: adding qualifiers `__host__ __device__` to the type transformer. 

It really weird that without those qualifiers,  there is runtime problem only when the vector size is 1.

Authors:
  - Nghia Truong (@ttnghia)

Approvers:
  - Mark Harris (@harrism)
  - @nvdbaranec
  - Conor Hoekstra (@codereport)
  - Vukasin Milovanovic (@vuule)

URL: https://github.com/rapidsai/cudf/pull/7382