Parquet reader:  Fix issue when using skip_rows on non-nested columns containing nulls (#7627)
Fixes  https://github.com/rapidsai/cudf/issues/7113

The basics of the fix are:  when reading non-nested columns, always start reading values at the beginning of the data stream and discarding ones that fall before skip_rows.  Previously we were attempting to jump directly to the first value to decode using a bad assumption.  This was possible with nested types because of the preprocess step we run, but not with flat/simple hierarchies.

Also contains some cleanup and const-ification of things.  In particular, I renamed the variable `out_pos` to `src_pos` in `page_state_s`.  The semantics of the variable are: "the position in the source data to read for the current output position", which made `out_pos` a super obscure name.

Authors:
  - @nvdbaranec

Approvers:
  - Devavret Makkar (@devavret)
  - Ashwin Srinath (@shwina)
  - Vukasin Milovanovic (@vuule)

URL: https://github.com/rapidsai/cudf/pull/7627