Fixing issue with explode_outer position not nulling position entries of null rows (#7754)
`explode_outer` supports writing a position column, but if the row was null it would incorrectly set the position to 0 and the row valid. Instead, it should null that position row as well. Luckily the null column matches 100% with the null column of the exploded column, so we can just copy it after it is created.

Fixes #7721

Authors:
  - Mike Wilson (@hyperbolic2346)

Approvers:
  - Conor Hoekstra (@codereport)
  - Jake Hemstad (@jrhemstad)

URL: https://github.com/rapidsai/cudf/pull/7754