Turn on NVTX by default in java build (#7761)
Investigation was done under https://github.com/NVIDIA/spark-rapids/issues/1721 and it showed no significant performance difference with NVTX on. It would make it a lot easier if this was on by default because it allows customers and developers to get trace with the same jar without having to go off and build a new CUDF version.

So this PR turns it on by default and adds in reading from environment variable if we need to change in the future from build scripts.

Authors:
  - Thomas Graves (@tgravescs)

Approvers:
  - Jason Lowe (@jlowe)
  - Robert (Bobby) Evans (@revans2)

URL: https://github.com/rapidsai/cudf/pull/7761