Replace OrderedDict with dict (#7847)
This PR replaces all uses of OrderedDict in core cudf files with raw dicts. Dicts are ordered as of Python 3.7, so we don't need OrderedDict, which is slower (and we're not utilizing `popitem` or `move_to_end`, its main extra methods). This PR also revealed some room for improving the `_grouped_names` attribute of `ColumnAccessor`. The `__getitem__` method of the new `_NestedGetItemDict` is ~6x faster than the old `NestedOrderedDict`.

This PR also revealed some deeper changes that might be possible in `ColumnAccessor`, but those require a bit more investigation and I'll put them in a subsequent PR.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Keith Kraus (https://github.com/kkraus14)

URL: https://github.com/rapidsai/cudf/pull/7847