Remove char_width member from cudf::string_view class (#7914)
This PR removes the `_char_width` member from the libcudf `cudf::string_view` class. This member was being used to record the character width of the bytes in the string but only if all the characters have the same width. This occurs when the string only contains ASCII encoded data which is all single-byte UTF-8 characters. The same optimization can be inferred when the existing `_length` and `_bytes` are equal. 
This change reduces the memory footprint of this class from 24 bytes to 16 bytes and therefore matches the size of `thrust::pair<char*,size_type>`. Using this class in a vector would thereby reduce the memory requirements for that vector by 1/3.

Authors:
  - David Wendt (https://github.com/davidwendt)

Approvers:
  - Vukasin Milovanovic (https://github.com/vuule)
  - Ram (Ramakrishna Prabhu) (https://github.com/rgsl888prabhu)
  - Mike Wilson (https://github.com/hyperbolic2346)

URL: https://github.com/rapidsai/cudf/pull/7914